chore(networks): rename coins-xrpl to coins-ripple
