+++ coins/coins-ripple/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@trezor/coins-xrpl",
+    "name": "@trezor/coins-ripple",
     "version": "10.0.0-alpha.1",
     "license": "MIT",
     "sideEffects": false,
