+++ coins/coins-ripple/src/constants/index.ts
