+++ coins/coins-ripple/src/types/index.ts
