+++ coins/coins-ripple/tsconfig.json
