+++ coins/coins-ripple/tsconfig.lib.json
