+++ packages/blockchain-link-utils/package.json
@@ -29,9 +29,9 @@
     },
     "dependencies": {
         "@noble/hashes": "^2.0.1",
+        "@trezor/coins-ripple": "workspace:*",
         "@trezor/coins-solana": "workspace:*",
         "@trezor/coins-stellar": "workspace:*",
-        "@trezor/coins-xrpl": "workspace:*",
         "@trezor/env-utils": "workspace:*",
         "@trezor/utils": "workspace:*"
     },
