+++ packages/blockchain-link-utils/tsconfig.json
@@ -2,9 +2,9 @@
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        { "path": "../../coins/coins-ripple" },
         { "path": "../../coins/coins-solana" },
         { "path": "../../coins/coins-stellar" },
-        { "path": "../../coins/coins-xrpl" },
         { "path": "../env-utils" },
         { "path": "../utils" },
         { "path": "../blockchain-link-types" },
