+++ packages/blockchain-link-utils/tsconfig.lib.json
@@ -6,9 +6,9 @@
         "rootDir": "./src"
     },
     "references": [
+        { "path": "../../coins/coins-ripple" },
         { "path": "../../coins/coins-solana" },
         { "path": "../../coins/coins-stellar" },
-        { "path": "../../coins/coins-xrpl" },
         { "path": "../env-utils" },
         { "path": "../utils" },
         { "path": "../blockchain-link-types" },
