+++ packages/blockchain-link/package.json
@@ -77,9 +77,9 @@
     "dependencies": {
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/blockchain-link-utils": "workspace:*",
+        "@trezor/coins-ripple": "workspace:*",
         "@trezor/coins-solana": "workspace:*",
         "@trezor/coins-stellar": "workspace:*",
-        "@trezor/coins-xrpl": "workspace:*",
         "@trezor/env-utils": "workspace:*",
         "@trezor/utils": "workspace:*",
         "@trezor/utxo-lib": "workspace:*",
