+++ packages/blockchain-link/tests/unit/connection.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable jest/no-jasmine-globals */
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { TimeoutError } from '@trezor/coins-xrpl';
+import { TimeoutError } from '@trezor/coins-ripple';
 import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
 import workers from './worker';
