+++ packages/blockchain-link/tsconfig.json
@@ -9,9 +9,9 @@
     "references": [
         { "path": "../blockchain-link-types" },
         { "path": "../blockchain-link-utils" },
+        { "path": "../../coins/coins-ripple" },
         { "path": "../../coins/coins-solana" },
         { "path": "../../coins/coins-stellar" },
-        { "path": "../../coins/coins-xrpl" },
         { "path": "../env-utils" },
         { "path": "../utils" },
         { "path": "../utxo-lib" },
