+++ packages/blockchain-link/tsconfig.lib.json
@@ -10,9 +10,9 @@
     "references": [
         { "path": "../blockchain-link-types" },
         { "path": "../blockchain-link-utils" },
+        { "path": "../../coins/coins-ripple" },
         { "path": "../../coins/coins-solana" },
         { "path": "../../coins/coins-stellar" },
-        { "path": "../../coins/coins-xrpl" },
         { "path": "../env-utils" },
         { "path": "../utils" },
         { "path": "../utxo-lib" },
