+++ packages/eslint/src/localRulesConfig.mjs
@@ -10,9 +10,9 @@ const publishableTrezorPackages = [
     '@trezor/blockchain-link-types',
     '@trezor/blockchain-link-utils',
     '@trezor/coins-cardano',
+    '@trezor/coins-ripple',
     '@trezor/coins-solana',
     '@trezor/coins-stellar',
-    '@trezor/coins-xrpl',
     '@trezor/connect',
     '@trezor/connect-common',
     '@trezor/connect-data',
